Extract primary metadata resolver
Move primary-image method identity, metadata judgments, caches, and narrow resolver adapters out of the assembly analysis builder while preserving its scheduling and aggregation role.
